refactor: define "f" in a file separate from app.ts "f" creates a circular dependency, because app.ts imports modules that depend on app.ts's "f". To solve this, I had moved the import statements in app.ts further down, which resolved the issue temporarily, but does not work with the way Vitest transforms the code (I assume it reorders import statements to always be at the top.)
