+++ src/app/app.ts
@@ -1,81 +1,5 @@
 import { InterfaceAdapter } from "../framework/app"
-import { buildF, BuiltArgument, Feature, fileSystemAutocompleter } from "../framework/features"
-
-export type X = {
-	FeatureContext: {
-		filen: FilenSDK,
-		cloudWorkingPath: CloudPath,
-	},
-	Feature: {
-		skipAuthentication: boolean,
-	},
-}
-const _f = buildF<X>()
-const cloudPath = ({ restrictType, skipCheckExists }: { restrictType?: "file" | "directory", skipCheckExists?: boolean }, arg: BuiltArgument<X, string | undefined>): BuiltArgument<X, CloudPath> => ({
-	spec: {
-		...arg.spec,
-		autocomplete: fileSystemAutocompleter({
-			restrictToDirectories: restrictType === "directory",
-			exists: async ({ x }, path) => {
-				try {
-					await x.filen.fs().stat({ path: x.cloudWorkingPath.navigate(path).toString() })
-					return true
-				} catch (e) {
-					if (e instanceof Error && e.name === "FileNotFoundError") return false
-					throw e
-				}
-			},
-			readdir: async ({ x }, pathStr) => {
-				const path = x.cloudWorkingPath.navigate(pathStr).toString()
-				// to get the list of items in a directory with type, first readdir() to populate the cache, and then access the internal cache directly
-				try {
-					await x.filen.fs().readdir({ path })
-					return Object.entries(x.filen.fs()._items)
-						.filter(([cachedPath]) => cachedPath.startsWith(path) && cachedPath !== path)
-						.map(([cachedPath, item]) => ({
-							name: cachedPath.includes("/") ? cachedPath.substring(cachedPath.lastIndexOf("/") + 1) : cachedPath,
-							isDirectory: item.type === "directory"
-						}))
-				} catch (e) {
-					if (e instanceof Error && e.name === "FileNotFoundError") return []
-					throw e
-				}
-			},
-			isDirectory: async ({ x }, path) => {
-				try {
-					const stat = await x.filen.fs().stat({ path: x.cloudWorkingPath.navigate(path).toString() })
-					return stat.type === "directory"
-				} catch (e) {
-					if (e instanceof Error && e.name === "FileNotFoundError") return false
-					throw e
-				}
-			}
-		})
-	},
-	value: async (ctx) => {
-		const path = ctx.x.cloudWorkingPath.navigate((await arg.value(ctx)) ?? "")
-		if (!skipCheckExists) {
-			const stat = await (async () => {
-				try {
-					return await ctx.x.filen.fs().stat({ path: path.toString() })
-				} catch (e) {
-					if (e instanceof Error && e.name === "FileNotFoundError") {
-						ctx.app.errExit(`No such cloud ${restrictType ?? "path"}: ${path.toString()}`)
-					}
-					throw e
-				}
-			})()
-			if (restrictType !== undefined && stat.type !== restrictType) {
-				ctx.app.errExit(`Not a ${restrictType}: ${path.toString()}`)
-			}
-		}
-		return path
-	}
-})
-export const f = { ..._f, cloudPath }
-
-// (these things can only be imported now, because they depend on `f`)
-
+import { Feature } from "../framework/features"
 import FilenSDK, { FilenSDKConfig } from "@filen/sdk"
 import path from "path"
 import os from "os"
@@ -90,6 +14,7 @@ import { syncCommand } from "./featureInterfaces/syncInterface"
 import { driveMountingCommand } from "./featureInterfaces/driveMountingInterface"
 import { webdavCommandGroup } from "./featureInterfaces/webdavInterface"
 import { s3Command } from "./featureInterfaces/s3Interface"
+import { f, X } from "./f"
 
 export const app = (argv: string[], adapter: InterfaceAdapter) => f.app({
 	info: {
