+++ src/test/tests.ts
@@ -3,7 +3,8 @@ import path from "path"
 import FilenSDK, { FilenSDKConfig } from "@filen/sdk"
 import fs from "fs/promises"
 import { rimraf } from "rimraf"
-import { app, X } from "../app/app"
+import { app } from "../app/app"
+import { X } from "../app/f"
 import { InterfaceAdapter } from "../framework/app"
 import { FeatureContext } from "../framework/features"
 import { CloudPath } from "../app/util/cloudPath"
